+++ grails-app/services/io/xh/hoist/http/BaseProxyService.groovy
@@ -2,7 +2,7 @@
  * This file belongs to Hoist, an application development toolkit
  * developed by Extremely Heavy Industries (www.xh.io | info@xh.io)
  *
- * Copyright © 2019 Extremely Heavy Industries Inc.
+ * Copyright © 2020 Extremely Heavy Industries Inc.
  */
 
 package io.xh.hoist.http
@@ -93,12 +93,10 @@ abstract class BaseProxyService extends BaseService {
     // Additional overrideable implementation methods
     //------------------------------------------------
     protected void installRequestHeaders(HttpServletRequest request, HttpRequestBase method) {
-        def names = proxyRequestHeaders()
-        Collection<String> send = request.headerNames.findAll {hasHeader(names, (String) it)}
+        def names = proxyRequestHeaders(),
+            send = (Collection<String>) request.headerNames.findAll { hasHeader(names, (String) it) }
 
-        send.each {
-            method.setHeader(it, request.getHeader(it))
-        }
+        send.each { method.setHeader(it, request.getHeader(it)) }
     }
 
     protected void installResponseHeaders(HttpServletResponse response, HttpResponse sourceResponse) {