+++ grails-app/services/io/xh/hoist/monitor/MonitoringService.groovy
@@ -2,7 +2,7 @@
  * This file belongs to Hoist, an application development toolkit
  * developed by Extremely Heavy Industries (www.xh.io | info@xh.io)
  *
- * Copyright © 2019 Extremely Heavy Industries Inc.
+ * Copyright © 2020 Extremely Heavy Industries Inc.
  */
 
 package io.xh.hoist.monitor
@@ -24,11 +24,14 @@ import static grails.util.Environment.isDevelopmentMode
 import static java.lang.System.currentTimeMillis
 
 /**
- * Service for application monitoring, governs generation of monitor results and status reports, analysis of these results,
- * and the emission of grails events to notify status conditions that may be of interest to the application.
+ * Coordinates application status monitoring. Requests monitor results and generates status reports
+ * on a configurable timer, analyzes the results, and publishes Grails events on status conditions
+ * of interest to the application.
  *
- * Monitor refresh timer and notification timer are disabled for local development.
- * If alertMode changes via forceRun in local development mode, notification will not be sent.
+ * In local development mode, auto-run/refresh of Monitors is disabled, but monitors can still be
+ * run on demand via forceRun(). Notification are never sent during local development.
+ *
+ * If enabled via config, this service will also write monitor run results to a dedicated log file.
  */
 class MonitoringService extends BaseService implements AsyncSupport, EventPublisher {
 
@@ -43,20 +46,17 @@ class MonitoringService extends BaseService implements AsyncSupport, EventPublis
     private Long _lastNotified
 
     void init() {
-        def monitorInterval = isDevelopmentMode() ? -1 : {monitorConfig.monitorRefreshMins}
         _monitorTimer = createTimer(
-                interval: monitorInterval,
-                intervalUnits: MINUTES,
-                delay: {monitorConfig.monitorStartupDelayMins},
-                delayUnits: MINUTES,
+                interval: {monitorInterval},
+                delay: startupDelay,
                 runFn: this.&onMonitorTimer
         )
 
-        def notifyInterval = isDevelopmentMode() ? -1 : 15
         _notifyTimer = createTimer (
-                interval: notifyInterval * SECONDS,
+                interval: {notifyInterval},
                 runFn: this.&onNotifyTimer
         )
+
         super.init()
     }
 
@@ -79,8 +79,10 @@ class MonitoringService extends BaseService implements AsyncSupport, EventPublis
     @Transactional
     private void runAllMonitors() {
         withDebug('Running monitors') {
+            def timeout = getTimeoutSeconds()
+
             def tasks = Monitor.list().collect { m ->
-                asyncTask { monitorResultService.runMonitor(m) }
+                asyncTask { monitorResultService.runMonitor(m, timeout) }
             }
 
             Map newResults = Promises
@@ -139,7 +141,7 @@ class MonitoringService extends BaseService implements AsyncSupport, EventPublis
     }
 
     private void notifyAlertModeChange() {
-        if(!isDevelopmentMode()) {
+        if (!isDevelopmentMode()) {
             notify('xhMonitorStatusReport', generateStatusReport())
             _lastNotified = currentTimeMillis()
         }
@@ -150,6 +152,7 @@ class MonitoringService extends BaseService implements AsyncSupport, EventPublis
 
         def failThreshold = monitorConfig.failNotifyThreshold,
             warnThreshold = monitorConfig.warnNotifyThreshold
+
         return _problems.values().any {
             it.cyclesAsFail >= failThreshold || it.cyclesAsWarn >= warnThreshold
         }
@@ -192,6 +195,23 @@ class MonitoringService extends BaseService implements AsyncSupport, EventPublis
         runAllMonitors()
     }
 
+    private int getMonitorInterval() {
+        return isDevelopmentMode() ? -1 : (monitorConfig.monitorRefreshMins * MINUTES)
+    }
+
+    private int getNotifyInterval() {
+        return isDevelopmentMode() ? -1 : (15 * SECONDS)
+    }
+
+    private int getStartupDelay() {
+        return monitorConfig.monitorStartupDelayMins * MINUTES
+    }
+
+    // Default supplied here as support for this sub-config was added late in the game.
+    private long getTimeoutSeconds() {
+        return monitorConfig.monitorTimeoutSecs ?: 15
+    }
+
     private JSONObject getMonitorConfig() {
         configService.getJSONObject('xhMonitorConfig')
     }
@@ -200,7 +220,8 @@ class MonitoringService extends BaseService implements AsyncSupport, EventPublis
         super.clearCaches()
         _results.clear()
         _problems.clear()
-        _monitorTimer.forceRun()
+        if (monitorInterval > 0) {
+            _monitorTimer.forceRun()
+        }
     }
-    
 }