+++ src/main/groovy/io/xh/hoist/exception/ExceptionRenderer.groovy
@@ -2,45 +2,56 @@
  * This file belongs to Hoist, an application development toolkit
  * developed by Extremely Heavy Industries (www.xh.io | info@xh.io)
  *
- * Copyright © 2019 Extremely Heavy Industries Inc.
+ * Copyright © 2020 Extremely Heavy Industries Inc.
  */
 
 package io.xh.hoist.exception
 
-import grails.gsp.PageRenderer
 import grails.util.GrailsUtil
 import groovy.transform.CompileStatic
 import groovy.util.logging.Slf4j
 import io.xh.hoist.json.JSONFormat
 import io.xh.hoist.json.JSONSerializer
-import io.xh.hoist.util.Utils
+import io.xh.hoist.log.LogSupport
 
 import javax.servlet.http.HttpServletRequest
 import javax.servlet.http.HttpServletResponse
 
 import static org.apache.http.HttpStatus.SC_FORBIDDEN
 import static org.apache.http.HttpStatus.SC_INTERNAL_SERVER_ERROR
+import static org.apache.http.HttpStatus.SC_NOT_FOUND
 
 @CompileStatic
 @Slf4j
 class ExceptionRenderer {
 
-    public template = null
-
     /**
-     * Main entry point.  Render an exception to the response.
+     * Main entry point.  Render an exception to the response and log appropriately.
      */
-    void render(Throwable t, HttpServletRequest request, HttpServletResponse response) {
+    void render(Throwable t, HttpServletRequest request, HttpServletResponse response, LogSupport logSupport = null) {
+
+        // Prepare exception
         GrailsUtil.deepSanitize(t)
-        response.setStatus(getHttpStatus(t))
-        switch (getResponseFormat(request)) {
-            case 'HTML':
-                renderAsHTML(t, request, response)
-                break
-            case 'JSON':
-                renderAsJSON(t, request, response)
-                break
+        int status = getHttpStatus(t)
+        if (t instanceof grails.validation.ValidationException) {
+            t = new ValidationException(t)
         }
+
+        // Log -- but don't trash logs for routine or client errors
+        if (logSupport) {
+            def message = t.message ?: 'Exception'
+            if (t instanceof RoutineException) {
+                logSupport.instanceLog.debug(message, t)
+            } else if (status < 500) {
+                logSupport.logErrorCompact(message, t)
+            } else {
+                logSupport.instanceLog.error(message, t)
+            }
+        }
+
+        // Render to response
+        response.setStatus(status)
+        renderAsJSON(t, request, response)
         response.flushBuffer()
     }
 
@@ -52,28 +63,10 @@ class ExceptionRenderer {
         response.writer.write(toJSON(t))
     }
 
-    protected void renderAsHTML(Throwable t, HttpServletRequest request, HttpServletResponse response) {
-        response.setContentType('text/html')
-        def isAuth = getHttpStatus(t) == SC_FORBIDDEN,
-            template = isAuth ? '/notAuthorized' : '/exception',
-            model = [
-                title: isAuth ? 'Not Authorized' : 'Exception',
-                exception: t,
-                exceptionAsJSON: toJSON(t)
-        ]
-
-        PageRenderer renderer = (PageRenderer) Utils.appContext.getBean('groovyPageRenderer')
-        String str = renderer.render(model: model, view: template)
-        response.writer.write(str)
-    }
-
     protected int getHttpStatus(Throwable t) {
         if (t instanceof NotAuthorizedException) return SC_FORBIDDEN
+        if (t instanceof NotFoundException) return SC_NOT_FOUND
         return SC_INTERNAL_SERVER_ERROR
-    }
-
-    protected String getResponseFormat(HttpServletRequest r) {
-        return r.getHeader('ACCEPT')?.contains('html') ? 'HTML' : 'JSON'
     }
 
     protected String toJSON(Throwable t) {
@@ -87,5 +80,4 @@ class ExceptionRenderer {
                 ].findAll {it.value}
         return JSONSerializer.serialize(ret);
     }
-    
 }