+++ src/main/groovy/io/xh/hoist/exception/ValidationException.groovy
@@ -1,0 +1,27 @@
+/*
+ * This file belongs to Hoist, an application development toolkit
+ * developed by Extremely Heavy Industries (www.xh.io | info@xh.io)
+ *
+ * Copyright © 2020 Extremely Heavy Industries Inc.
+ */
+
+package io.xh.hoist.exception
+
+import groovy.transform.CompileStatic
+import io.xh.hoist.util.Utils
+
+/**
+ * Exception for use when user input fails GORM validation
+ */
+class ValidationException extends RuntimeException implements RoutineException {
+    ValidationException(grails.validation.ValidationException ex) {
+        super(parseMessage(ex), ex)
+    }
+
+    private static String parseMessage(grails.validation.ValidationException ex) {
+        def msgSrc = Utils.appContext.messageSource
+        return ex.errors.allErrors
+                .collect {msgSrc.getMessage(it, Locale.US)}
+                .join(' | ')
+    }
+}